+++ modules/tinymce/CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Color picker dialog would not update the preview color if the hex input value was prefixed with `#` symbol. #TINY-9457
 - Table cell selection was possible even if the element was noneditable in a noneditable root element. #TINY-9459
 - Table commands were modifying noneditable tables. #TINY-9459
+- Fake carets were rendered for noneditable element and tables if the root was noneditable. #TINY-9459
 
 ## 6.3.1 - 2022-12-06